added changes specifically for deploying to Azure app servivce or container app
