+++ CHANGELOG
@@ -1,25 +1,5 @@
-Version 1.0.5 [14-01-2022]
-    - Fixed Chrome Webstore downloader (issue #13)
-    - Added support for local google chrome extensions for Mac Os
-    - Added support for Vivaldi Browser (issue #20)
-    - Minor bug fixes
-Version 1.0.4 [14-05-2020]
-    - Fixed an error where local firefox extensions could not be detected (issue #9)
-    - Added support for Brave browser
-    - Fixed some CSRF issues
-    - Added `--nobrowser` argument
-Version 1.0.3 [22-10-2019]
-    - Show extension browser name in scan info (issue #4)
-    - Fixed a bug that created error when user is offline
-Version 1.0.2 [14-05-2019]
-    - Removed dummy virustotal column
-    - Show "vulnerabilities" button along js file only if any vulnerabilities are detected
-    - Important files check
-    - Frontend scan option
-Version 1.0.1 [09-05-2019]
-    - No Javascript Warning added
-    - Auto create lab directory if not exist
-    - Fix Chrome Web Store downloader
-    - Updater fixed for windows
-Version 1.0.0 [06-05-2019]
-    - Initial Launch
+Version 1.0.1 [11-10-2023]
+    - Changed ExtAnalysis infrastructure from Azure Web App to Azure Container App, which fixed the permhash library import error of ```raise importerror('failed to find libmagic. check your installation')```
+Version 1.0.0 [10-10-2023]
+    - Added permhash in ExtAnalysis ```requirements.txt``` and use the library to calculate permhash value
+    - Added permhash column in SQL Database
