+++ templates/report.html
@@ -129,7 +129,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
     </div>
     <br><br>
     <div class="tabscontainer">
-        <meta name="csrf-token" content="{{ csrf_token() }}">
 
       <ul class="result-tabs">
         <li class="tab-link current" data-tab="tab-0"><i class="fas fa-tree"></i> Basic Info</li>
